# app/routes/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Designed to be consumed by the Integrated Pipeline Reports (IPR) Client, the endpoints are\navailable for authorized users to consume.",
    "version": "1.0.0",
    "title": "Integrated Pipeline Reports (IPR)",
    "contact": {
      "name": "IPR Team",
      "email": "ipr@bcgsc.ca"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "paths": {
...
    "/germline-small-mutation-reports/{germline}/reviews": {
      "get": {
        "summary": "Get Germline Reviews",
        "description": "Retrieve all reviews for a given germline report",
        "parameters": [
          {
            "$ref": "#/components/parameters/germline"
          }
        ],
        "tags": [
          "Germline Reviews"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of reviews for the given germline report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/germlineSmallMutationReviewAssociations"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report not found"
          }
        }
      }
    },
    "/germline-small-mutation-reports/{germline}/reviews/{review}": {
      "get": {
        "summary": "Get Germline Review",
        "description": "Retrieve the specified review for a given germline report",
        "parameters": [
          {
            "$ref": "#/components/parameters/germline"
          },
          {
            "$ref": "#/components/parameters/review"
          }
        ],
        "tags": [
          "Germline Reviews"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the details of the specified review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/germlineSmallMutationReviewAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report or review not found"
          }
        }
      },
      "put": {
        "summary": "Update Germline Review",
        "description": "Update the details of a germline review",
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          },
          {
            "$ref": "#/components/parameters/review"
          }
        ],
        "requestBody": {
          "description": "New values to update current germline review",
          "required": true,
          "content": {
            "application/json":
            {
              "schema": {
                "$ref": "#/components/schemas/germlineSmallMutationReviewUpdate"
              }
            }
          }
        },
        "tags": [
          "Germline Reviews"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the updated germline review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/germlineSmallMutationReviewAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report or review not found"
          }
        }
      },
      "delete": {
        "summary": "Delete Germline Review",
        "description": "Removes the specified germline review",
        "parameters": [
          {
            "$ref": "#/components/parameters/germline"
          },
          {
            "$ref": "#/components/parameters/review"
          }
        ],
        "tags": [
          "Germline Reviews"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report or review not found"
          }
        }
      }
    },
    "/germline-small-mutation-reports/{germline}/variants": {
      "get": {
        "summary": "Get Germline Variants",
        "description": "Retrieve all variants for a given germline report",
        "parameters": [
          {
            "$ref": "#/components/parameters/germline"
          }
        ],
        "tags": [
          "Germline Variants"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of variants for the given germline report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/germlineSmallMutationVariantAssociations"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report not found"
          }
        }
      }
    },
    "/germline-small-mutation-reports/{germline}/variants/{variant}": {
      "get": {
        "summary": "Get Germline Variant",
        "description": "Retrieve the specified variant for a given germline report",
        "parameters": [
          {
            "$ref": "#/components/parameters/germline"
          },
          {
            "$ref": "#/components/parameters/variant"
          }
        ],
        "tags": [
          "Germline Variants"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the details of the specified variant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/germlineSmallMutationVariantAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report or variant not found"
          }
        }
      },
      "put": {
        "summary": "Update Germline Variant",
        "description": "Update the details of a germline variant",
        "parameters": [
          {
            "$ref": "#/components/parameters/germline"
          },
          {
            "$ref": "#/components/parameters/variant"
          }
        ],
        "requestBody": {
          "description": "New values to update current germline variant",
          "required": true,
          "content": {
            "application/json":
            {
              "schema": {
                "$ref": "#/components/schemas/germlineSmallMutationVariantUpdate"
              }
            }
          }
        },
        "tags": [
          "Germline Variants"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the updated germline variant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/germlineSmallMutationVariantAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report or variant not found"
          }
        }
      },
      "delete": {
        "summary": "Delete Germline Variant",
        "description": "Removes the specified germline variant",
        "parameters": [
          {
            "$ref": "#/components/parameters/germline"
          },
          {
            "$ref": "#/components/parameters/variant"
          }
        ],
        "tags": [
          "Germline Variants"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Germline report or variant not found"
          }
        }
      }
    },
...
    "/reports/{report}/small-mutations": {
      "get": {
        "summary": "Get Small Mutations",
        "description": "Retrieve all small mutations for a given report",
        "tags": [
          "Small Mutations"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of somatic small mutations for the given report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/smallMutationsAssociations"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Report not found"
          }
        }
      }
    },
    "/reports/{report}/small-mutations/{mutation}": {
      "get": {
        "summary": "Get Small Mutation",
        "description": "Retrieves a small mutation from the report by ident",
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          },
          {
            "$ref": "#/components/parameters/mutation"
          }
        ],
        "tags": [
          "Small Mutations"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the details of the specified small mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/smallMutationsAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Report or small mutation not found"
          }
        }
      },
      "put": {
        "summary": "Update Small Mutation",
        "description": "Update the details of a small mutation",
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          },
          {
            "$ref": "#/components/parameters/mutation"
          }
        ],
        "requestBody": {
          "description": "New values to update current small mutation",
          "required": true,
          "content": {
            "application/json":
            {
              "schema": {
                "$ref": "#/components/schemas/smallMutationsUpdate"
              }
            }
          }
        },
        "tags": [
          "Small Mutations"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the details of the updated small mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/smallMutationsAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Report or small mutation not found"
          }
        }
      },
      "delete": {
        "summary": "Delete Small Mutation",
        "description": "Removes the specified small mutation",
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          },
          {
            "$ref": "#/components/parameters/mutation"
          }
        ],
        "tags": [
          "Small Mutations"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Report or small mutation not found"
          }
        }
      }
    },
...
  },
...
  "components": {
    "parameters": {
      "report": {
        "name": "report",
        "in": "path",
        "required": true,
        "description": "Analysis report ident",
        "schema": {
          "type": "string",
          "format": "UUIDv4"
        }
      },
      "germline": {
        "name": "germline",
        "in": "path",
        "required": true,
        "description": "Germline report ident",
        "schema": {
          "type": "string",
          "format": "UUIDv4"
        }
      },
...
      "review": {
        "name": "review",
        "in": "path",
        "required": true,
        "description": "Germline review ident",
        "schema": {
          "type": "string",
          "format": "UUIDv4"
        }
      },
      "variant": {
        "name": "variant",
        "in": "path",
        "required": true,
        "description": "Germline variant ident",
        "schema": {
          "type": "string",
          "format": "UUIDv4"
        }
      },
...
      "mutation": {
        "name": "mutation",
        "in": "path",
        "required": true,
        "description": "small mutation ident",
        "schema": {
          "type": "string",
          "format": "UUIDv4"
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "headers": {
          "WWW_Authenticate": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}